# Span-ify PathMarkupParser