- What does the pull request do?
Change `PathMarkupParser` to use `ReadOnlySpan<char>`s instead of regular expressions to parse paths.
- What is the current behavior?
Currently the parser uses regular expressions and data-structures on the heap to parse paths.
- What is the updated/expected behavior with this PR?
Better performance. 

Old perf (from #1661):

``` ini

BenchmarkDotNet=v0.10.8, OS=Windows 10.0.17134
Processor=Intel Core i7-4500U CPU 1.80GHz (Haswell), ProcessorCount=4
Frequency=2338342 Hz, Resolution=427.6534 ns, Timer=TSC
dotnet cli version=2.1.300
  [Host]     : .NET Core 4.6.26328.01, 64bit RyuJIT
  DefaultJob : .NET Core 4.6.26328.01, 64bit RyuJIT


```
 |           Method |     Mean |    Error |   StdDev |   Gen 0 |  Gen 1 | Allocated |
 |----------------- |---------:|---------:|---------:|--------:|-------:|----------:|
 | Parse_Large_Path | 171.5 us | 3.389 us | 5.936 us | 18.3105 | 2.4414 |  37.66 KB |

New perf:

``` ini

BenchmarkDotNet=v0.10.14, OS=Windows 10.0.17134
Intel Core i5-6300U CPU 2.40GHz (Skylake), 1 CPU, 4 logical and 2 physical cores
.NET Core SDK=2.1.300
  [Host]     : .NET Core 2.0.7 (CoreCLR 4.6.26328.01, CoreFX 4.6.26403.03), 64bit RyuJIT
  DefaultJob : .NET Core 2.0.7 (CoreCLR 4.6.26328.01, CoreFX 4.6.26403.03), 64bit RyuJIT


```
|           Method |     Mean |     Error |    StdDev |  Gen 0 |  Gen 1 | Allocated |
|----------------- |---------:|----------:|----------:|-------:|-------:|----------:|
| Parse_Large_Path | 29.60 us | 0.5703 us | 0.5602 us | 0.8850 | 0.2136 |   4.76 KB |


So, this PR uses approximately 13% of the managed memory that the current implementation uses in our benchmark.