@KvanTTT 2018-06-23
Did you measure performance on different processors? I see `Intel Core i7-4500U CPU 1.80GHz (Haswell), ProcessorCount=4` in old perf and `Intel Core i5-6300U CPU 2.40GHz (Skylake), 1 CPU, 4 logical and 2 physical cores` in the new one.