@jkoritzinsky 2018-06-25
I don't believe so. The android project build would be failing if it was blocked.